LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/navigation-within-beforeunload-expected.txt


FAIL Triggering navigation from within beforeunload event assert_equals: expected 1000 but got 1
